Partial factor out of common trace code.
