+++ apis/Google.Cloud.Diagnostics.AspNet/Google.Cloud.Diagnostics.AspNet.Tests/Trace/TraceHeaderTraceOptionsFactoryTest.cs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.using Microsoft.VisualStudio.TestTools.UnitTesting;
 
+using Google.Cloud.Diagnostics.Common;
 using Xunit;
 
 namespace Google.Cloud.Diagnostics.AspNet.Tests
