+++ apis/Google.Cloud.Diagnostics.Common/Google.Cloud.Diagnostics.Common.Tests/Trace/RateLimitingTraceOptionsFactoryTest.cs
@@ -14,7 +14,7 @@
 
 using Xunit;
 
-namespace Google.Cloud.Diagnostics.AspNet.Tests
+namespace Google.Cloud.Diagnostics.Common.Tests
 {
     public class RateLimitingTraceOptionsFactoryTest
     {
