+++ apis/Google.Cloud.Diagnostics.Common/Google.Cloud.Diagnostics.Common/project.json
@@ -18,12 +18,18 @@
   },
 
   "dependencies": {
+    "Google.Api.CommonProtos": "1.0.0-beta05",
     "Google.Api.Gax": "1.0.0-beta06",
+    "Google.Cloud.Trace.V1": { "target": "project" }
   },
 
 
   "frameworks": {
     "net45": {},
-    "netstandard1.5": {}
+    "netstandard1.5": {
+      "dependencies": {
+        "System.Diagnostics.StackTrace": "4.3.0"
+      }
+    }
   }
 }
